requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
celery==5.3.4
redis==5.0.1
djangorestframework-simplejwt==5.3.0
django-environ==0.11.2
Pillow>=10.0.0
opencv-python>=4.8.0
openai>=1.0.0
gTTS>=2.4.0
plotly>=5.17.0
pandas>=2.1.0
numpy>=1.24.0
scikit-learn>=1.3.0
django-extensions>=3.2.0
gunicorn>=21.0.0
whitenoise>=6.6.0
python-decouple>=3.8
channels>=4.0.0
channels-redis>=4.1.0
daphne>=4.0.0
django-filter>=23.5